+++ packages/sdk-react/CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.0]
+### Uncategorized
+- feat: rpc read only calls and infura provider ([#353](https://github.com/MetaMask/metamask-sdk/pull/353))
+- feat: add codecov to CI ([#343](https://github.com/MetaMask/metamask-sdk/pull/343))
+
 ## [0.6.2]
 ### Uncategorized
 - docs: add readme info for sdk-react and sdk-react-ui ([#342](https://github.com/MetaMask/metamask-sdk/pull/342))
@@ -39,7 +44,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [fix] publishing config ([#135](https://github.com/MetaMask/metamask-sdk/pull/135))
 - [feat] initial beta released
 
-[Unreleased]: https://github.com/MetaMask/metamask-sdk/compare/v0.6.2...HEAD
+[Unreleased]: https://github.com/MetaMask/metamask-sdk/compare/v0.7.0...HEAD
+[0.7.0]: https://github.com/MetaMask/metamask-sdk/compare/v0.6.2...v0.7.0
 [0.6.2]: https://github.com/MetaMask/metamask-sdk/compare/v0.6.1...v0.6.2
 [0.6.1]: https://github.com/MetaMask/metamask-sdk/compare/v0.6.0...v0.6.1
 [0.6.0]: https://github.com/MetaMask/metamask-sdk/compare/v0.5.6...v0.6.0